man: Man page reference cleanup
Sphinx is stricter than pandoc when parsing reST and does not allow an
opening parenthesis to directly follow inline markup. [1] describes the
rules. A backslash, as described in [2], can be used to separate the
two.

Additionally, there were different forms for referring to man pages.
Sometimes it was a plain “ganeti(7)”, sometimes “**ganeti**(7)” and
other times “**ganeti(7)**” (of course with different names and
numbers).

Commit 7acbda7b added a new paragraph and used the Sphinx-specific
“:manpage:…” markup, which isn't supported by pandoc. It is replaced
with a standard reference.

[1]
http://docutils.sourceforge.net/docs/ref/rst/restructuredtext.html#
inline-markup-recognition-rules

[2]
http://docutils.sourceforge.net/docs/ref/rst/restructuredtext.html#
character-level-inline-markup

Signed-off-by: Michael Hanselmann <hansmi@google.com>
Reviewed-by: Guido Trotter <ultrotter@google.com>